added authorisation to the basket resource. Created nested objects basket with shared_with users

+++ baskets/views.py
@@ -1,9 +1,11 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
 from .serializers.common import BasketSerializer
 from .models import Basket
 from rest_framework.permissions import IsAuthenticated
+from utils.permissions import IsOwnerOrShared
+from .serializers.populate import PopulatedBasketSerializer
 
 # Create your views here.
 class HelloWorldView (APIView):
@@ -26,6 +28,7 @@ class BasketsView (APIView):
         return Response (serializer.data, status=201)
 
 class BasketsDetailsView(APIView):
+    permission_classes = [IsOwnerOrShared]    
     
     def get_object (self, pk): 
         try: 
@@ -35,10 +38,10 @@ class BasketsDetailsView(APIView):
     
     def get (self, request, pk):
         basket = self.get_object(pk)
-        serializer = BasketSerializer(basket)
+        serializer =PopulatedBasketSerializer(basket)
         return Response (serializer.data)
     
-    def put (self, request, pk):
+    def put (self, request, pk):      
         basket = self.get_object (pk)
         serializer = BasketSerializer(basket, data= request.data, partial = True)
         serializer.is_valid(raise_exception = True)
@@ -46,6 +49,12 @@ class BasketsDetailsView(APIView):
         return Response (serializer.data)
         
     def delete (self, request, pk):
+        basket= self.get_object (pk)
+        
+        #basic ownership check       
+        if basket.owner != request.user:
+            raise PermissionDenied (f'Only {basket.owner} may delete this basket.')
+        
         basket = self.get_object(pk)
         basket.delete()
         return Response (status = 204)
